requirements.txt
python-dotenv
langchain
transformers
sentence-transformers
streamlit
langchain_experimental
langchain_openai
huggingface-hub
langchain-community
langchain-core
langchain-experimental
langchain-openai
langsmith
mysql
mysqlclient
nltk
openai
pinecone-client
google-auth  
regex
psycopg2-binary
pkgconfig
PyPDF2
oauth2client
sqlalchemy
pandas